+++ runtime-testsuite/test/org/antlr/v5/test/runtime/kotlin/commonTest/org/antlr/v5/runtime/kotlin/CharStreamsTest.kt
